enable responsiveness
Tapping on the inputs on mobile causes a zoom to them, so that after you
have selected a new city (for example), you're still all zoomed in on just
the name of the city, rather than returning to the previous zoom. In addition,
I couldn't find a javascript way to re-zoom, short of triggering a full page
reload.